[nested-containers] - Pretty sure everything works for 1 level of nesting, now it needs to be made somewhat recursive to support multiple layers of nesting and we are probably good - TT
