+++ Tests/Swift/NestedPrivateExtensionClassTest.swift
@@ -1,0 +1,8 @@
+//
+//  NestedPrivateExtensionClassTest.swift
+//  Cuckoo
+//
+//  Created by thompsty on 9/18/20.
+//
+
+import Foundation
