+++ Tests/Swift/Source/NestedInPrivateNestedClass.swift
@@ -1,0 +1,14 @@
+//
+//  NestedInPrivateNestedClass.swift
+//  Cuckoo
+//
+//  Created by thompsty on 9/18/20.
+//
+
+import Foundation
+
+fileprivate class PrivateNestedClass {
+    class ThisClassShouldNotBeMocked2 {
+        var property: Int?
+    }
+}
