+++ Tests/Swift/Source/NestedPrivateExtensionClass.swift
@@ -1,0 +1,14 @@
+//
+//  NestedPrivateExtensionClass.swift
+//  Cuckoo
+//
+//  Created by thompsty on 9/18/20.
+//
+
+import Foundation
+
+fileprivate extension Nested {
+    class ThisClassShouldNotBeMocked1 {
+        var property: Int?
+    }
+}
